“get assignment status”

+++ platform-pojo/src/main/java/team8/ad/project/entity/entity/AssignmentDetails.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class AssignmentDetails {
-    private Long id;
-    private Long assignmentId;
-    private Long questionId;
+    private Integer id;
+    private Integer assignmentId;
+    private Integer questionId;
 }
